+++ tests/functional/adapter/incremental/fixtures.py
@@ -762,6 +762,22 @@ models:
         columns: [id]
 """
 
+schema_with_single_column_primary_key_constraint_removed = """
+version: 2
+
+models:
+  - name: primary_key_constraint_sql
+    columns:
+      - name: id
+        data_type: bigint
+        constraints:
+          - type: not_null
+      - name: version
+        data_type: int
+      - name: msg
+        data_type: string
+"""
+
 schema_with_composite_primary_key_constraint = """
 version: 2
 
@@ -1010,4 +1026,44 @@ def model(dbt, spark):
     )
     data = [[1, 'hello', 'blue']]
     return spark.createDataFrame(data, schema=['id', 'msg', 'color'])
+"""
+
+warn_unenforced_override_sql = """
+select "abc" as id
+"""
+
+warn_unenforced_override_model = """
+version: 2
+
+models:
+  - name: model_a
+    config:
+      materialized: incremental
+      on_schema_change: fail
+    columns:
+      - name: id
+        data_type: string
+        constraints:
+          - type: not_null
+    constraints:
+      - type: primary_key
+        columns:
+          - id
+        name: model_a_pk
+        warn_unenforced: false
+
+  - name: model_b
+    config:
+        materialized: table
+    columns:
+      - name: id
+        data_type: string
+    constraints:
+      - type: foreign_key
+        columns:
+          - id
+        name: model_b_fk
+        to: ref('model_a')
+        to_columns: [id]
+        warn_unenforced: false
 """